feat(diary): AI 요청 util 생성
- AI 요청 테스트하기 위한 util입니다.
- 현재 localhost로 하드코딩되어 있어 다음 회의 때 설정파일 추가하면 될 것 같습니다.
